feat(connect): refactor InspectorModal component
