+++ apps/connect/src/components/modal/modals/InspectorModal/index.ts
@@ -1,0 +1,1 @@
+export { InspectorModal } from "./InspectorModal.js";
