+++ apps/connect/src/components/modal/modals/InspectorModal/useDbExplorer.ts
@@ -1,16 +1,7 @@
 import { pgDump } from "@electric-sql/pglite-tools/pg_dump";
-import { InspectorModal as ConnectInspectorModal } from "@powerhousedao/design-system/connect";
-import { closePHModal, usePHModal } from "@powerhousedao/reactor-browser";
-import {
-  useDatabase,
-  usePGlite,
-  useSync,
-} from "@powerhousedao/reactor-browser/connect";
-
+import { useDatabase, usePGlite } from "@powerhousedao/reactor-browser/connect";
 import { sql } from "kysely";
 import { useCallback } from "react";
-
-const DEFAULT_PAGE_SIZE = 25;
 
 type ColumnInfo = {
   readonly name: string;
@@ -48,12 +39,9 @@ type TablePage = {
   readonly total: number | null;
 };
 
-export const InspectorModal: React.FC = () => {
-  const phModal = usePHModal();
+export function useDbExplorer() {
   const database = useDatabase();
   const pglite = usePGlite();
-  const syncManager = useSync();
-  const open = phModal?.type === "inspector";
 
   const getTables = useCallback(async (): Promise<TableInfo[]> => {
     if (!database) return [];
@@ -73,7 +61,6 @@ export const InspectorModal: React.FC = () => {
       ORDER BY c.table_name, c.ordinal_position
     `.execute(database);
 
-    // Group columns by table name
     const tableMap = new Map<string, ColumnInfo[]>();
     for (const row of result.rows) {
       if (!tableMap.has(row.table_name)) {
@@ -86,7 +73,6 @@ export const InspectorModal: React.FC = () => {
       });
     }
 
-    // Convert map to array of TableInfo
     return Array.from(tableMap).map(([name, columns]) => ({
       name,
       columns,
@@ -95,16 +81,13 @@ export const InspectorModal: React.FC = () => {
 
   const getTableRows = useCallback(
     async (table: string, options: GetTableRowsOptions): Promise<TablePage> => {
-      if (!database || !pglite) {
+      if (!database) {
         return { columns: [], rows: [], total: null };
       }
 
       const { limit, offset, sort } = options;
-
-      // Use explicit schema prefix to avoid search_path issues
       const tableRef = sql.raw(`public."${table}"`);
 
-      // Build the query with optional sorting
       let query;
       if (sort) {
         const columnRef = sql.raw(`"${sort.column}"`);
@@ -123,7 +106,6 @@ export const InspectorModal: React.FC = () => {
 
       const result = await query.execute(database);
 
-      // Get total count
       const countResult = await sql<{ count: string }>`
         SELECT COUNT(*) as count FROM ${tableRef}
       `.execute(database);
@@ -131,7 +113,6 @@ export const InspectorModal: React.FC = () => {
         ? parseInt(countResult.rows[0].count, 10)
         : null;
 
-      // Extract column names from first row or return empty
       const columns = result.rows.length > 0 ? Object.keys(result.rows[0]) : [];
 
       return {
@@ -162,40 +143,18 @@ export const InspectorModal: React.FC = () => {
     async (sqlContent: string) => {
       if (!pglite) return;
 
-      // Drop and recreate public schema to clear all tables
       await pglite.exec(`DROP SCHEMA public CASCADE`);
       await pglite.exec(`CREATE SCHEMA public`);
-
-      // Execute the import
       await pglite.exec(sqlContent);
-
-      // Reset search_path to public after import (pg_dump may change it)
       await pglite.exec(`SET search_path TO public`);
     },
     [pglite],
   );
 
-  const getRemotes = useCallback(() => {
-    return Promise.resolve(syncManager?.list() ?? []);
-  }, [syncManager]);
-
-  return (
-    <ConnectInspectorModal
-      open={open}
-      onOpenChange={(status) => {
-        if (!status) closePHModal();
-      }}
-      dbExplorerProps={{
-        schema: "public",
-        getTables,
-        getTableRows,
-        pageSize: DEFAULT_PAGE_SIZE,
-        onExportDb,
-        onImportDb,
-      }}
-      remotesInspectorProps={{
-        getRemotes,
-      }}
-    />
-  );
-};
+  return {
+    getTables,
+    getTableRows,
+    onExportDb,
+    onImportDb,
+  };
+}
