+++ apps/connect/src/components/modal/modals/InspectorModal/useRemotesInspector.ts
@@ -1,0 +1,14 @@
+import { useSync } from "@powerhousedao/reactor-browser/connect";
+import { useCallback } from "react";
+
+export function useRemotesInspector() {
+  const syncManager = useSync();
+
+  const getRemotes = useCallback(() => {
+    return Promise.resolve(syncManager?.list() ?? []);
+  }, [syncManager]);
+
+  return {
+    getRemotes,
+  };
+}
